+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [4.9.1](https://github.com/plivo/plivo-go/tree/v4.9.1) (2020-08-10)
+- Fix Get Details of a Call API response.
+
+## [4.9.0](https://github.com/plivo/plivo-go/tree/v4.9.0) (2020-08-04)
+- Add service type support(SMS/MMS) for Powerpack API.
+
+## [4.8.0](https://github.com/plivo/plivo-go/tree/v4.8.0) (2020-07-23)
+- Add retries to multiple regions for voice requests.
+
+## [4.7.1](https://github.com/plivo/plivo-go/tree/v4.7.1) (2020-07-13)
+- Fix Call Create & Retrieve Call details API responses.
+
 ## [4.7.0](https://github.com/plivo/plivo-go/tree/v4.7.0) (2020-05-28)
 - Add JWT helper functions.